sandbox-image/tool-api/requirements.txt
fastapi
uvicorn
python-pty
requests
beautifulsoup4
googlesearch-python
psutil
